client/src/utils/requirements.txt
Pillow==10.2.0
pillow-heif==0.13.0 